fixed up in the mvc structure, lots of things left before this branch is finished
