--- src/main/java/edu/ntnu/boardgame/actions/ChessPuzzleAction.java
@@ -1,10 +1,0 @@
-package edu.ntnu.boardgame.actions;
-
-import edu.ntnu.boardgame.gui.ChessPuzzleView;
-
-public class ChessPuzzleAction {
-
-  public void launchPuzzle() {
-    new ChessPuzzleView().show();
-  }
-}
