+++ src/main/java/edu/ntnu/boardgame/actions/tileactions/TeleportRandomAction.java
@@ -1,4 +1,4 @@
-package edu.ntnu.boardgame.actions;
+package edu.ntnu.boardgame.actions.tileactions;
 
 import edu.ntnu.boardgame.Board;
 import edu.ntnu.boardgame.constructors.Player;
